% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_tools.R
\name{install_bioc_if_missing}
\alias{install_bioc_if_missing}
\title{Install Bioconductor Package if Missing}
\usage{
install_bioc_if_missing(package)
}
\arguments{
\item{package}{The name of the package to check and install.}
}
\description{
This function checks if a package is installed from Bioconductor and installs it if not.
It uses the `BiocManager` package to install the package from the specified Bioconductor repository.
}
