% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_tools.R
\name{install_github_if_missing}
\alias{install_github_if_missing}
\title{Install GitHub Package if Missing}
\usage{
install_github_if_missing(repo)
}
\arguments{
\item{repo}{The GitHub repository in the format "username/repo".}
}
\description{
This function checks if a package is installed from GitHub and installs it if not.
It uses the `devtools` package to install the package from a specified GitHub repository.
}
