% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_tools.R
\name{install_if_missing}
\alias{install_if_missing}
\title{Conditionally Install Packages}
\usage{
install_if_missing(package, host = "CRAN", repo = NULL)
}
\arguments{
\item{package}{The name of the package to check and install.}

\item{host}{The host from which to install the package. Options are "CRAN", "GitHub", "R-Forge", or "Bioconductor".}

\item{repo}{The GitHub repository in the format "username/repo" (for GitHub installations).}
}
\description{
This function checks if a package is installed and installs it if not.
It also provides options for installing from GitHub, R-Forge, or Bioconductor.
}
