% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_tools.R
\name{install_rforge_if_missing}
\alias{install_rforge_if_missing}
\title{Install R-Forge Package if Missing}
\usage{
install_rforge_if_missing(package, repo = "http://R-Forge.R-project.org")
}
\arguments{
\item{package}{The name of the package to check and install.}

\item{repo}{The R-Forge repository URL. Default is "http://R-Forge.R-project.org".}
}
\description{
This function checks if a package is installed from R-Forge and installs it if not.
It uses the `install.packages` function to install the package from the specified R-Forge repository.
}
