% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sample_no_surprises}
\alias{sample_no_surprises}
\title{Function to handle single or multiple samples}
\usage{
sample_no_surprises(x)
}
\arguments{
\item{x}{A vector from which to sample.}
}
\value{
A single element from the vector `x`.
}
\description{
This function takes a vector `x` and returns a single random sample from it.
If the vector has only one element, it returns that element without attempting to sample.
This prevents errors that can occur when trying to sample from a single-element vector.
}
